Admin now adds the user, activation email is send and user sets up the password. Added password reset option

+++ website/views.py
@@ -38,19 +38,22 @@ def posts_list(request):
 
 @login_required
 def post_draft_list(request):
-    template = 'website/post_draft_list.html'
-    post_list = Post.objects.filter(status='Draft').order_by('created')
-    page = request.GET.get('page', 1)
+    if request.user.is_staff:
+        template = 'website/post_draft_list.html'
+        post_list = Post.objects.filter(status='Draft').order_by('created')
+        page = request.GET.get('page', 1)
 
-    paginator = Paginator(post_list, 6)
-    try:
-        posts = paginator.page(page)
-    except PageNotAnInteger:
-        posts = paginator.page(1)
-    except EmptyPage:
-        posts = paginator.page(paginator.num_pages)
+        paginator = Paginator(post_list, 6)
+        try:
+            posts = paginator.page(page)
+        except PageNotAnInteger:
+            posts = paginator.page(1)
+        except EmptyPage:
+            posts = paginator.page(paginator.num_pages)
 
-    return render(request, template, {'posts': posts})
+        return render(request, template, {'posts': posts})
+    else:
+        raise Http404
 
 
 def post_detail(request, pk):
@@ -60,30 +63,33 @@ def post_detail(request, pk):
 
 @login_required
 def post_new(request):
-    template = 'website/post_edit.html'
-    if request.method == "POST":
-        form = PostForm(request.POST)
+    if request.user.is_staff:
+        template = 'website/post_edit.html'
+        if request.method == "POST":
+            form = PostForm(request.POST)
 
-        try:
-            if form.is_valid():
-                if 'cancel' in request.POST:
-                    return HttpResponseRedirect(get_success_url())
-                else:
-                    post = form.save(commit=False)
-                    post.author = request.user
-                    post.save()
-                    return redirect('website:post_detail', pk=post.pk)
+            try:
+                if form.is_valid():
+                    if 'cancel' in request.POST:
+                        return HttpResponseRedirect(get_success_url())
+                    else:
+                        post = form.save(commit=False)
+                        post.author = request.user
+                        post.save()
+                        return redirect('website:post_detail', pk=post.pk)
 
-        except Exception as e:
-            messages.warning(request, 'Post Failed To Save. Error: {}".format(e)')
+            except Exception as e:
+                messages.warning(request, 'Post Failed To Save. Error: {}".format(e)')
 
+        else:
+            form = PostForm()
+        context = {
+            'form': form,
+        }
+
+        return render(request, template, context)
     else:
-        form = PostForm()
-    context = {
-        'form': form,
-    }
-
-    return render(request, template, context)
+        raise Http404
 
 
 def get_success_url():
@@ -92,11 +98,9 @@ def get_success_url():
 
 @login_required
 def post_edit(request, pk):  # also post update
-    template = 'website/post_edit.html'
-    post = get_object_or_404(Post, pk=pk)
-
-    # Only author of the post can edit the post
-    if request.user == post.author:
+    if request.user.is_staff:
+        template = 'website/post_edit.html'
+        post = get_object_or_404(Post, pk=pk)
         # If this is a POST request then process the Form data
         if request.method == "POST":
 
@@ -125,43 +129,51 @@ def post_edit(request, pk):  # also post update
         }
 
         return render(request, template, context)
+    else:
+        raise Http404
 
 
 @login_required
 def post_publish(request, pk):
-    post = get_object_or_404(Post, pk=pk)
-    template = 'website/confirmation_publish.html'
+    if request.user.is_staff:
+        post = get_object_or_404(Post, pk=pk)
+        template = 'website/confirmation_publish.html'
 
-    if request.method == "POST":
-        post.publish()
-        messages.success(request, "This post has been published.")
-        return HttpResponseRedirect(reverse('website:posts_list'))
+        if request.method == "POST":
+            post.publish()
+            messages.success(request, "This post has been published.")
+            return HttpResponseRedirect(reverse('website:posts_list'))
 
-    if request.user != post.author:
-        raise PermissionDenied
+        if request.user != post.author:
+            raise PermissionDenied
 
-    context = {
-        "post": post
-    }
+        context = {
+            "post": post
+        }
 
-    return render(request, template, context)
+        return render(request, template, context)
+    else:
+        raise Http404
 
 
 @login_required
 def post_remove(request, pk):
-    post = get_object_or_404(Post, pk=pk)
-    template = 'website/confirmation_delete.html'
+    if request.user.is_staff:
+        post = get_object_or_404(Post, pk=pk)
+        template = 'website/confirmation_delete.html'
 
-    if request.method == "POST":
-        post.delete()
-        messages.success(request, "This has been deleted.")
-        return HttpResponseRedirect(reverse('website:posts_list'))
+        if request.method == "POST":
+            post.delete()
+            messages.success(request, "This has been deleted.")
+            return HttpResponseRedirect(reverse('website:posts_list'))
 
-    if request.user != post.author:
-        raise PermissionDenied
+        if request.user != post.author:
+            raise PermissionDenied
 
-    context = {
-        "post": post
-    }
+        context = {
+            "post": post
+        }
 
-    return render(request, template, context)
+        return render(request, template, context)
+    else:
+        raise Http404
